Custom hooks both for adding and deleting todos and showing errors were added
